middleware updates- implemented middleware on most endpoints
