+++ apis/external/auth0.js
@@ -42,7 +42,6 @@ function getUserProfile(accessToken) {
             return res.data;
         })
         .catch(err => {
-            console.log(err)
             console.log('Error retrieving user profile information from Auth0', err.message);
         });
 }
