+++ config/users/router.js
@@ -1,16 +1,16 @@
 const express = require('express');
-const usersDb = require('../apis/db/users');
+const usersDb = require('../../apis/db/users');
 
 const router = express.Router();
 
 router.get('/', (req, res) => {
 
-    //Filter out sub_id since it's already available client-side
+    //Filter out sub_id since it is not needed client-side
     const user = req.headers.user;
     delete user.sub_id;
 
     //Return user information
-    //It is already populated from the middleware
+    //It is already populated from the auth middleware
     res.status(200).send(user);
 });
